license-plate-removal/web-service/requirements.txt
requests
streamlit
pillow